+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -69,6 +69,7 @@ class RichContentEditor extends Component {
     this.deprecateSiteDomain();
     this.initContext();
     this.initPlugins();
+    setTimeout(() => this.publish(), 3000);
   }
 
   componentDidUpdate() {